requirements.txt
pytorch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio==0.8.0
Pillow==9.5.0
timm==0.3.2
lifelines==0.27.4